requirements.txt
openai
flask
python-dotenv
posthog
